Combined source image and processed image into one.
