--- .vscode/launch.json
@@ -1,12 +1,0 @@
-{
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "type": "node",
-      "request": "launch",
-      "name": "Launch Program",
-      "program": "${workspaceRoot}/cli.js",
-      "args": ["--overwrite", "paintface.jpg"]
-    }
-  ]
-}
