--- .vscode/settings.json
@@ -1,6 +1,0 @@
-{
-  "cSpell.words": [
-    "Quantizer",
-    "colodore"
-  ]
-}
