--- retropixels.code-workspace
@@ -1,20 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "web"
-		},
-		{
-			"path": "cli"
-		},
-		{
-			"path": "core"
-		}
-	],
-	"settings": {
-		"cSpell.words": [
-			"Quantizer",
-			"colodore",
-			"quantizer"
-		]
-	}
-}
